moved send egld/esdt tests from basic-features to forwarder
