+++ contracts/feature-tests/async/mandos/message_otherShard.scen.json
@@ -1,5 +1,6 @@
 {
     "name": "message_otherShard",
+    "gasSchedule": "dummy",
     "steps": [
         {
             "step": "setState",
