+++ contracts/feature-tests/async/mandos/message_otherShard_callback.scen.json
@@ -1,5 +1,6 @@
 {
     "name": "message_otherShard_callback",
+    "gasSchedule": "dummy",
     "steps": [
         {
             "step": "setState",
