+++ contracts/feature-tests/async/mandos/message_sameShard_callback.scen.json
@@ -1,5 +1,6 @@
 {
     "name": "message_sameShard_callback",
+    "gasSchedule": "dummy",
     "steps": [
         {
             "step": "setState",
