+++ contracts/feature-tests/async/mandos/payment_otherShard.scen.json
@@ -1,5 +1,6 @@
 {
     "name": "payment_otherShard",
+    "gasSchedule": "dummy",
     "steps": [
         {
             "step": "setState",
