+++ contracts/feature-tests/async/mandos/payment_otherShard_callback.scen.json
@@ -1,5 +1,6 @@
 {
     "name": "payment_otherShard_callback",
+    "gasSchedule": "dummy",
     "steps": [
         {
             "step": "setState",
