+++ contracts/feature-tests/async/mandos/payment_sameShard.scen.json
@@ -1,5 +1,6 @@
 {
     "name": "payment_sameShard",
+    "gasSchedule": "dummy",
     "steps": [
         {
             "step": "setState",
