+++ contracts/feature-tests/async/mandos/send_esdt.scen.json
@@ -1,5 +1,5 @@
 {
-    "name": "send_egld",
+    "name": "send_esdt",
     "gasSchedule": "v2",
     "steps": [
         {
@@ -7,9 +7,12 @@
             "accounts": {
                 "address:features_contract": {
                     "nonce": "1000",
-                    "balance": "200",
+                    "balance": "0",
                     "storage": {},
-                    "code": "file:../output/basic-features.wasm"
+                    "esdt": {
+                        "str:BASIC-FEATURES-TOKEN": "1,000"
+                    },
+                    "code": "file:../forwarder/output/forwarder.wasm"
                 },
                 "address:an_account": {
                     "nonce": "0",
@@ -26,9 +29,10 @@
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "function": "send_egld",
+                "function": "send_esdt",
                 "arguments": [
                     "address:an_account",
+                    "str:BASIC-FEATURES-TOKEN",
                     "10000"
                 ],
                 "gasLimit": "50,000,000",
@@ -37,7 +41,7 @@
             "expect": {
                 "out": [],
                 "status": "*",
-                "message": "str:failed transfer (insufficient funds)",
+                "message": "str:insufficient funds",
                 "logs": []
             }
         },
@@ -48,9 +52,10 @@
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "function": "send_egld",
+                "function": "send_esdt",
                 "arguments": [
                     "address:an_account",
+                    "str:BASIC-FEATURES-TOKEN",
                     "100"
                 ],
                 "gasLimit": "50,000,000",
@@ -69,9 +74,10 @@
                 "from": "address:an_account",
                 "to": "address:features_contract",
                 "value": "0",
-                "function": "send_egld",
+                "function": "send_esdt",
                 "arguments": [
                     "address:an_account",
+                    "str:BASIC-FEATURES-TOKEN",
                     "100",
                     "str:send tx message here"
                 ],
@@ -91,12 +97,18 @@
                     "nonce": "1000",
                     "balance": "0",
                     "storage": {},
-                    "code": "file:../output/basic-features.wasm"
+                    "esdt": {
+                        "str:BASIC-FEATURES-TOKEN": "800"
+                    },
+                    "code": "file:../forwarder/output/forwarder.wasm"
                 },
                 "address:an_account": {
                     "nonce": "*",
-                    "balance": "200",
+                    "balance": "0",
                     "storage": {},
+                    "esdt": {
+                        "str:BASIC-FEATURES-TOKEN": "200"
+                    },
                     "code": ""
                 }
             }
