+++ contracts/feature-tests/basic-features/tests/basic_features_test.rs
@@ -1,4 +1,3 @@
-extern crate basic_features;
 use basic_features::*;
 use elrond_wasm::types::{SCError, SCResult};
 use elrond_wasm_debug::*;
